% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_obs_grid.R
\name{is_obs_grid}
\alias{is_obs_grid}
\title{Test for flux_obs_grid}
\usage{
is_obs_grid(x)
}
\arguments{
\item{x}{Object.}
}
\value{
Logical.
}
\description{
Return \code{TRUE} if object inherits from \code{flux_obs_grid}.
}
